+++ src/remotedebug/rdebug_cfunctioninfo.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <optional>
+#include <string>
+
+namespace remotedebug {
+	std::optional<std::string> get_functioninfo(void* ptr);
+}
